refactor(ips): move price description component
ref: #MANAGER-17130
