+++ packages/manager/apps/ips/public/translations/ips/Messages_fr_FR.json
@@ -1,6 +1,5 @@
 {
   "breadcrumb_root_label": "Toutes les IPs",
   "order": "Commander une IP",
-  "free_price": "Gratuit",
-  "per_ip": "/IP"
+  "free_price": "Gratuit"
 }
